third_party/blink/web_tests/accessibility/svg-image-expected.txt

 

This tests that SVG images are accessible elements and they have accessible names as real images.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svgImage.role is realImage.role
PASS svgImage.name is realImage.name
SVG Image Role: AXRole: AXImage
SVG Image Description: TestImage
PASS successfullyParsed is true

TEST COMPLETE
